+++ core/testsuite/src/test/java/com/blazebit/persistence/testsuite/PaginationEmbeddedIdTest.java
@@ -175,12 +175,18 @@ public class PaginationEmbeddedIdTest extends AbstractCoreTest {
         String expectedObjectQuery = "SELECT e FROM EmbeddableTestEntity e"
                 + " WHERE (e.id.key = :ids_0_0 AND e.id.value = :ids_1_0)"
                 + " ORDER BY e.id.key ASC, e.id.value ASC";
+        String expectedInlineObjectQuery = "SELECT e, e.id.key, e.id.value FROM EmbeddableTestEntity e"
+                + " WHERE " + function("compare_row_value_subquery", "'IN'", "e.id.key", "e.id.value", function("LIMIT", "(" + expectedIdQuery + ")", "1")) + " = true"
+                + " ORDER BY e.id.key ASC, e.id.value ASC";
 
         PaginatedCriteriaBuilder<EmbeddableTestEntity> pcb = crit.page(null, 0, 1);
 
         assertEquals(expectedCountQuery, pcb.getPageCountQueryString());
-        assertEquals(expectedIdQuery, pcb.getPageIdQueryString());
+        assertEquals(expectedIdQuery, pcb.withInlineIdQuery(false).getPageIdQueryString());
         assertEquals(expectedObjectQuery, pcb.getQueryString());
+        if (dbmsDialect.supportsRowValueConstructor()) {
+            assertEquals(expectedInlineObjectQuery, pcb.withInlineIdQuery(true).getQueryString());
+        }
         PagedList<EmbeddableTestEntity> resultList = pcb.getResultList();
 
         pcb = crit.page(resultList.getKeysetPage(), 1, 1);
@@ -199,10 +205,16 @@ public class PaginationEmbeddedIdTest extends AbstractCoreTest {
         expectedObjectQuery = "SELECT e FROM EmbeddableTestEntity e"
                 + " WHERE (e.id.key = :ids_0_0 AND e.id.value = :ids_1_0)"
                 + " ORDER BY e.id.key ASC, e.id.value ASC";
+        expectedInlineObjectQuery = "SELECT e, e.id.key, e.id.value FROM EmbeddableTestEntity e"
+                + " WHERE " + function("compare_row_value_subquery", "'IN'", "e.id.key", "e.id.value", function("LIMIT", "(" + expectedIdQuery + ")", "1")) + " = true"
+                + " ORDER BY e.id.key ASC, e.id.value ASC";
 
         assertEquals(expectedCountQuery, pcb.getPageCountQueryString());
-        assertEquals(expectedIdQuery, pcb.getPageIdQueryString());
+        assertEquals(expectedIdQuery, pcb.withInlineIdQuery(false).getPageIdQueryString());
         assertEquals(expectedObjectQuery, pcb.getQueryString());
+        if (dbmsDialect.supportsRowValueConstructor()) {
+            assertEquals(expectedInlineObjectQuery, pcb.withInlineIdQuery(true).getQueryString());
+        }
         pcb.getResultList();
     }
 
@@ -221,11 +233,18 @@ public class PaginationEmbeddedIdTest extends AbstractCoreTest {
         crit.orderByAsc("e.id.key");
         crit.orderByAsc("e.id.value");
 
+        String expectedIdQuery = "SELECT e.id.key, e.id.value FROM EmbeddableTestEntity e LEFT JOIN e.embeddable.elementCollection elementCollection_1 WHERE elementCollection_1.primaryName = :param_0 GROUP BY e.id.key, e.id.value ORDER BY e.id.key ASC, e.id.value ASC";
         String expectedObjectQuery = "SELECT e FROM EmbeddableTestEntity e"
                 + " WHERE (e.id.key = :ids_0_0 AND e.id.value = :ids_1_0)"
                 + " ORDER BY e.id.key ASC, e.id.value ASC";
-        String queryString = crit.page(0, 1).getQueryString();
-        assertEquals(expectedObjectQuery, queryString);
+        String expectedInlineObjectQuery = "SELECT e FROM EmbeddableTestEntity e"
+                + " WHERE " + function("compare_row_value_subquery", "'IN'", "e.id.key", "e.id.value", function("LIMIT", "(" + expectedIdQuery + ")", "1")) + " = true"
+                + " ORDER BY e.id.key ASC, e.id.value ASC";
+        PaginatedCriteriaBuilder<EmbeddableTestEntity> pcb = crit.page(0, 1);
+        assertEquals(expectedObjectQuery, pcb.withInlineIdQuery(false).getQueryString());
+        if (dbmsDialect.supportsRowValueConstructor()) {
+            assertEquals(expectedInlineObjectQuery, pcb.withInlineIdQuery(true).getQueryString());
+        }
     }
 
     @Test
@@ -239,11 +258,17 @@ public class PaginationEmbeddedIdTest extends AbstractCoreTest {
         crit.orderByAsc("e.id.value");
         PaginatedCriteriaBuilder<Tuple> pcb = crit.pageBy(0, 1, "e.id.key", "e.id.value");
 
+        String expectedIdQuery = "SELECT e.id.key, e.id.value FROM EmbeddableTestEntity e WHERE e.id.key = :param_0 GROUP BY e.id.key, e.id.value ORDER BY e.id.key ASC, e.id.value ASC";
         String expectedObjectQuery = "SELECT e.id, elementCollection_1.primaryName FROM EmbeddableTestEntity e LEFT JOIN e.embeddable.elementCollection elementCollection_1"
                 + " WHERE (e.id.key = :ids_0_0 AND e.id.value = :ids_1_0)"
                 + " ORDER BY e.id.key ASC, e.id.value ASC";
-        String queryString = pcb.getQueryString();
-        assertEquals(expectedObjectQuery, queryString);
+        String expectedInlineObjectQuery = "SELECT e.id, elementCollection_1.primaryName FROM EmbeddableTestEntity e LEFT JOIN e.embeddable.elementCollection elementCollection_1"
+                + " WHERE " + function("compare_row_value_subquery", "'IN'", "e.id.key", "e.id.value", function("LIMIT", "(" + expectedIdQuery + ")", "1")) + " = true"
+                + " ORDER BY e.id.key ASC, e.id.value ASC";
+        assertEquals(expectedObjectQuery, pcb.withInlineIdQuery(false).getQueryString());
+        if (dbmsDialect.supportsRowValueConstructor()) {
+            assertEquals(expectedInlineObjectQuery, pcb.withInlineIdQuery(true).getQueryString());
+        }
         PagedList<Tuple> list = pcb.getResultList();
         assertEquals(2, list.size());
     }