+++ integration/openjpa/src/main/java/com/blazebit/persistence/integration/openjpa/OpenJPAJpaProvider.java
@@ -196,6 +196,21 @@ public class OpenJPAJpaProvider implements JpaProvider {
     }
 
     @Override
+    public boolean supportsNonScalarSubquery() {
+        return false;
+    }
+
+    @Override
+    public boolean supportsSubqueryInFunction() {
+        return false;
+    }
+
+    @Override
+    public boolean supportsSubqueryAliasShadowing() {
+        return true;
+    }
+
+    @Override
     public String[] getDiscriminatorColumnCheck(EntityType<?> entityType) {
         return null;
     }